完成ng-table的使用，显示图册表格，具有列表显示、排序、根据图册名称查询功能

+++ src/main/java/com/hc/dao/AuthorityRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthorityRepository extends CrudRepository<Authority, Integer> {
+public interface AuthorityRepository extends CrudRepository<Authority, Long> {
 
 }
